+++ modules/apps/depot/depot-test/src/testIntegration/java/com/liferay/depot/internal/change/tracking/spi/reference/test/DepotEntryGroupRelTableReferenceDefinitionTest.java
@@ -3,10 +3,12 @@
  * SPDX-License-Identifier: LGPL-2.1-or-later OR LicenseRef-Liferay-DXP-EULA-2.0.0-2023-06
  */
 
-package com.liferay.depot.change.tracking.test;
+package com.liferay.depot.internal.change.tracking.spi.reference.test;
 
 import com.liferay.arquillian.extension.junit.bridge.junit.Arquillian;
 import com.liferay.change.tracking.test.util.BaseTableReferenceDefinitionTestCase;
+import com.liferay.depot.model.DepotEntry;
+import com.liferay.depot.service.DepotEntryGroupRelLocalService;
 import com.liferay.depot.service.DepotEntryLocalService;
 import com.liferay.portal.kernel.model.change.tracking.CTModel;
 import com.liferay.portal.kernel.test.rule.AggregateTestRule;
@@ -20,6 +22,7 @@ import com.liferay.portal.test.rule.PermissionCheckerMethodTestRule;
 
 import java.util.HashMap;
 
+import org.junit.Before;
 import org.junit.ClassRule;
 import org.junit.Rule;
 import org.junit.runner.RunWith;
@@ -28,7 +31,7 @@ import org.junit.runner.RunWith;
  * @author Brooke Dalton
  */
 @RunWith(Arquillian.class)
-public class DepotEntryTableReferenceDefinitionTest
+public class DepotEntryGroupRelTableReferenceDefinitionTest
 	extends BaseTableReferenceDefinitionTestCase {
 
 	@ClassRule
@@ -38,14 +41,28 @@ public class DepotEntryTableReferenceDefinitionTest
 			new LiferayIntegrationTestRule(),
 			PermissionCheckerMethodTestRule.INSTANCE);
 
+	@Before
 	@Override
-	protected CTModel<?> addCTModel() throws Exception {
-		return _depotEntryLocalService.addDepotEntry(
+	public void setUp() throws Exception {
+		super.setUp();
+
+		_depotEntry = _depotEntryLocalService.addDepotEntry(
 			HashMapBuilder.put(
 				LocaleUtil.getDefault(), RandomTestUtil.randomString()
 			).build(),
 			new HashMap<>(), ServiceContextTestUtil.getServiceContext());
 	}
+
+	@Override
+	protected CTModel<?> addCTModel() throws Exception {
+		return _depotEntryGroupRelLocalService.addDepotEntryGroupRel(
+			_depotEntry.getDepotEntryId(), group.getGroupId());
+	}
+
+	private DepotEntry _depotEntry;
+
+	@Inject
+	private DepotEntryGroupRelLocalService _depotEntryGroupRelLocalService;
 
 	@Inject
 	private DepotEntryLocalService _depotEntryLocalService;
