+++ modules/apps/depot/depot-test/src/testIntegration/java/com/liferay/depot/internal/change/tracking/spi/reference/test/DepotEntryTableReferenceDefinitionTest.java
@@ -3,14 +3,11 @@
  * SPDX-License-Identifier: LGPL-2.1-or-later OR LicenseRef-Liferay-DXP-EULA-2.0.0-2023-06
  */
 
-package com.liferay.depot.change.tracking.test;
+package com.liferay.depot.internal.change.tracking.spi.reference.test;
 
 import com.liferay.arquillian.extension.junit.bridge.junit.Arquillian;
 import com.liferay.change.tracking.test.util.BaseTableReferenceDefinitionTestCase;
-import com.liferay.depot.model.DepotEntry;
-import com.liferay.depot.service.DepotAppCustomizationLocalService;
 import com.liferay.depot.service.DepotEntryLocalService;
-import com.liferay.journal.constants.JournalPortletKeys;
 import com.liferay.portal.kernel.model.change.tracking.CTModel;
 import com.liferay.portal.kernel.test.rule.AggregateTestRule;
 import com.liferay.portal.kernel.test.util.RandomTestUtil;
@@ -23,7 +20,6 @@ import com.liferay.portal.test.rule.PermissionCheckerMethodTestRule;
 
 import java.util.HashMap;
 
-import org.junit.Before;
 import org.junit.ClassRule;
 import org.junit.Rule;
 import org.junit.runner.RunWith;
@@ -32,7 +28,7 @@ import org.junit.runner.RunWith;
  * @author Brooke Dalton
  */
 @RunWith(Arquillian.class)
-public class DepotAppCustomizationTableReferenceDefinitionTest
+public class DepotEntryTableReferenceDefinitionTest
 	extends BaseTableReferenceDefinitionTestCase {
 
 	@ClassRule
@@ -42,29 +38,14 @@ public class DepotAppCustomizationTableReferenceDefinitionTest
 			new LiferayIntegrationTestRule(),
 			PermissionCheckerMethodTestRule.INSTANCE);
 
-	@Before
 	@Override
-	public void setUp() throws Exception {
-		super.setUp();
-
-		_depotEntry = _depotEntryLocalService.addDepotEntry(
+	protected CTModel<?> addCTModel() throws Exception {
+		return _depotEntryLocalService.addDepotEntry(
 			HashMapBuilder.put(
 				LocaleUtil.getDefault(), RandomTestUtil.randomString()
 			).build(),
 			new HashMap<>(), ServiceContextTestUtil.getServiceContext());
 	}
-
-	@Override
-	protected CTModel<?> addCTModel() throws Exception {
-		return _depotAppCustomizationLocalService.updateDepotAppCustomization(
-			_depotEntry.getDepotEntryId(), false, JournalPortletKeys.JOURNAL);
-	}
-
-	@Inject
-	private DepotAppCustomizationLocalService
-		_depotAppCustomizationLocalService;
-
-	private DepotEntry _depotEntry;
 
 	@Inject
 	private DepotEntryLocalService _depotEntryLocalService;
